backend/app/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
sqlmodel
openai
python-multipart
pdfplumber
pydantic